changing the int64 to integer (just to keep the same standard in all columns)

+++ part2/1. ddl.sql
@@ -4,9 +4,9 @@ create schema if not exists `bounce-challenge`.ecomm;
 -- 1. `orders` (columns: `order_id`, `customer_id`, `product_id`, `order_time`, `quantity`)
 create table if not exists ecomm.orders
 (
-  order_id int64,
-  customer_id int64,
-  product_id int64,
+  order_id integer,
+  customer_id integer,
+  product_id integer,
   order_time datetime,
   quantity integer
 );
@@ -14,7 +14,7 @@ create table if not exists ecomm.orders
 -- 2. `products` (columns: `product_id`, `product_name`, `category`, `price`)
 create table if not exists ecomm.products
 (
-  product_id int64,
+  product_id integer,
   product_name string,
   category string,
   price float64
@@ -23,7 +23,7 @@ create table if not exists ecomm.products
 -- 3. `customers` (columns: `customer_id`, `registration_date`, `country`)
 create table if not exists ecomm.customers
 (
-  customer_id int64,
+  customer_id integer,
   registration_date date,
   country string
 );
